Add files via upload
This repository contains a stand-alone version of the code used to analyze the transition in sniping behaviour among high-frequency traders. It produces the solution in which probabilistic sniping is more profitable.  In order to turn the solution into  a  viable  strategy  for  the  repeated  game,  we  show  how  compliance can be monitored through the use of sequential statistical testing
If you make use of this code please cite our work: Kokabisaghi, Pauwels, Dorsman, (2020).
The package of our code covers the simulation study and Numerical results. We also referred each section of the codes in our paper and vice versa.  
Code can be run using python assuming the following packages are installed (most are defaults in every python installation):
numpy
pandas
matplotlib
sympy